modified gnl to return the length of the return string

+++ includes/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: kdavis <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/28 11:09:03 by kdavis            #+#    #+#             */
-/*   Updated: 2016/12/21 13:39:27 by kdavis           ###   ########.fr       */
+/*   Updated: 2017/01/03 18:51:31 by kdavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # define BUFF_SIZE 1024
 
-int		get_next_line(const int fd, char **line);
+ssize_t	get_next_line(const int fd, char **line);
 int		ft_getll_base(char **str, long long *nbr);
 
 #endif
